# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Overtime Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /up-next:
    get:
      operationId: getUpNext
      tags: [reports]
      summary: Get ordered overtime rotation for a roster
      parameters:
        - name: rosterId
          in: query
          required: true
          schema:
            type: integer
        - name: dayType
          in: query
          required: false
          schema:
            type: string
            default: weekday
      responses:
        "200":
          description: Ordered rotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpNextResult"
  /stats:
    get:
      operationId: getStats
      tags: [reports]
      summary: Get summary statistics
      parameters:
        - name: rosterId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
  /suggest-day-type:
    get:
      operationId: suggestDayType
      tags: [reports]
      summary: Suggest a day type for a given date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Suggested day type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayTypeSuggestion"
components:
  schemas:
...
    Event:
      type: object
      required: [id, rosterId, date, description, defaultHours, dayType, multiplier, entries]
      properties:
        id:
          type: integer
        rosterId:
          type: integer
        date:
          type: string
          format: date
        description:
          type: string
        defaultHours:
          type: number
        dayType:
          type: string
        multiplier:
          type: number
          default: 1
        entries:
          type: array
          items:
            $ref: "#/components/schemas/EventEntry"
    EventEntry:
      type: object
      required: [id, employeeId, employeeName, offered, worked, hoursOffered, hoursAwarded]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        offered:
          type: boolean
        worked:
          type: boolean
        hoursOverride:
          type: ["number", "null"]
        hoursAwarded:
          type: number
        hoursOffered:
          type: number
...
    UpNextResult:
      type: object
      required: [rosterId, dayType, employees]
      properties:
        rosterId:
          type: integer
        dayType:
          type: string
        employees:
          type: array
          items:
            $ref: "#/components/schemas/UpNextEmployee"
    UpNextEmployee:
      type: object
      required: [id, name, seniority, totalOfferedHours, fairnessScore, rank]
      properties:
        id:
          type: integer
        name:
          type: string
        subclassId:
          type: ["integer", "null"]
        subclassName:
          type: ["string", "null"]
        roleName:
          type: ["string", "null"]
        seniority:
          type: integer
        totalOfferedHours:
          type: number
        fairnessScore:
          type: number
        rank:
          type: integer
...
    Stats:
      type: object
      required: [totalEvents, totalOfferedHours, totalWorkedHours, employeeCount]
      properties:
        totalEvents:
          type: integer
        totalOfferedHours:
          type: number
        totalWorkedHours:
          type: number
        employeeCount:
          type: integer
        topWorkers:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeStat"
        recentEvents:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        normalizedHoursResetEventId:
          type: ["integer", "null"]
          description: ID of the most recent "Reset Hours" event used for time-window filtering
    EmployeeStat:
      type: object
      required: [id, name, totalWorkedHours, totalOfferedHours]
      properties:
        id:
          type: integer
        name:
          type: string
        totalWorkedHours:
          type: number
        totalOfferedHours:
          type: number
    DayTypeSuggestion:
      type: object
      required: [date, suggestedDayType, reason]
      properties:
        date:
          type: string
          format: date
        suggestedDayType:
          type: string
          enum: [weekday, weekend, holiday]
        reason:
          type: string
        isHoliday:
          type: boolean
        holidayName:
          type: ["string", "null"]