# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Overtime Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeStat } from './employeeStat';
import type { Event } from './event';

export interface Stats {
  totalEvents: number;
  totalOfferedHours: number;
  totalWorkedHours: number;
  employeeCount: number;
  topWorkers?: EmployeeStat[];
  recentEvents?: Event[];
  /**
     * ID of the most recent "Reset Hours" event used for time-window filtering
     * @nullable
     */
  normalizedHoursResetEventId?: number | null;
}